Changed number data types from float to decimal for precision

+++ db/migrate/20220112161902_create_transactions.rb
@@ -3,11 +3,11 @@ class CreateTransactions < ActiveRecord::Migration[6.0]
     create_table :transactions do |t|
       t.string :kind
       t.string :symbol
-      t.float :unit_price
+      t.decimal :unit_price
       t.decimal :amount, :precision => 15, :scale => 10
-      t.float :total_price
-      t.float :fee
-      t.float :final_price
+      t.decimal :total_price
+      t.decimal :fee
+      t.decimal :final_price
       t.string :currency
       t.string :status
       t.integer :user_id
